+++ main/server/fixtures.js
@@ -1,5 +1,5 @@
-import { Lists } from '../api/lists/lists.js';
-import { Todos } from '../api/todos/todos.js';
+import { Lists } from '../../imports/api/lists/lists.js';
+import { Todos } from '../../imports/api/todos/todos.js';
 
 // if the database is empty on server start, create some sample data.
 Meteor.startup(() => {
