+++ main/server/register-api.js
@@ -1,2 +1,2 @@
 // This module defines all methods and publications on the server
-import '../imports/api/api.js';
+import '../../imports/api/api.js';
